Refactored injected scripts to their own files
