+++ source/scripts/ContentScript/index.ts
@@ -1,26 +1,11 @@
-
 import { providerManager, stargazerProvider } from './inject';
 
 import { Script } from 'scripts/Provider/Script';
 
 new Script().start();
 
-inject(providerManager());
-inject(stargazerProvider());
-
-// function injectEthereum (asset: string, name: string) {
-//
-//   inject(ethereumProvider({
-//     name,
-//     asset,
-//     network: {
-//       networkId: 1, chainId: 1
-//     },
-//     overrideEthereum: true
-//   }))
-// }
-//
-// injectEthereum('ETH', 'eth')
+inject(providerManager);
+inject(stargazerProvider);
 
 function inject(content: string) {
   const container = document.head || document.documentElement;
