+++ source/scripts/ContentScript/inject/index.ts
@@ -1,0 +1,6 @@
+// @ts-nocheck
+// Read files in as strings
+import stargazerProvider from './stargazerProvider.txt';
+import providerManager from './providerManager.txt';
+
+export { providerManager, stargazerProvider };
