+++ source/web/package.json
@@ -137,6 +137,7 @@
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "postcss-loader": "^3.0.0",
     "prettier": "^2.0.5",
+    "raw-loader": "^4.0.2",
     "resolve-url-loader": "^3.1.1",
     "sass-loader": "^8.0.2",
     "style-loader": "^1.2.1",
